\name{d1W_1}
\alias{d1W_1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{First derivative of W (non-principal branch)}
\description{
Computes the first derivative of the Lambert W function (principal branch).
}
\usage{
d1W_1(z)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{z}{  value \code{z} where the derivative shall be computed}
}

\value{
Derivative of \eqn{W_1(.)} evaluated at z.
}
\references{Goerg, G.M. (2009). \dQuote{Lambert W Random Variables - A new class of skewed distribution functions}. Unpublished}
\author{Georg M. Goerg}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{math}
