\name{delta.test}
\alias{delta.test}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Test for significant delta}
\description{
Given the simulated critical values (see References) for delta =0, a test on the significance is performed on the estimate of delta.
}
\usage{
delta.test(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{object of class \code{LWest}}
  \item{...}{...}
}

\value{
p-value of the test statistic
}
\references{Goerg, G.M. (2009). \dQuote{Lambert W Random Variables - A new class of skewed distribution functions}. Unpublished}
\author{Georg M. Goerg}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{htest}
