% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{deprecated-functions}
\alias{deprecated-functions}
\alias{beta_names}
\alias{bounds_theta}
\alias{d1W_1}
\alias{p_1}
\alias{params2theta}
\alias{skewness_test}
\alias{starting_theta}
\alias{support}
\alias{normfit}
\alias{theta2params}
\alias{vec.norm}
\alias{W_1}
\alias{W_gamma_1}
\alias{H}
\title{List of deprecated functions}
\usage{
beta_names(...)

bounds_theta(...)

d1W_1(z, W.z = W(z, branch = -1))

p_1(...)

params2theta(...)

skewness_test(...)

starting_theta(...)

support(...)

normfit(...)

theta2params(...)

vec.norm(...)

W_1(z)

W_gamma_1(z, gamma)

H(...)
}
\arguments{
\item{...}{arguments passed to deprecated functions.}

\item{z, W.z}{see \code{\link{deriv_W}}}

\item{gamma}{see \code{\link{W_gamma}}.}
}
\description{
These functions have been deprecated in v0.5 of \pkg{LambertW} mostly for
    sake of following R style guides with respect to naming of
    functions. This means that all deprecated functions here have an
    analogous function with a similar -- more style consistent -- name. See
    also the \code{NEWS} file.

Deprecated functions still work as expected, but they print out a warning
suggesting to use the new function (name).
}
