% Generated by roxygen2 (4.0.2): do not edit by hand
\name{distname-utils}
\alias{check_distname}
\alias{distname-utils}
\alias{get_distname_family}
\alias{get_distnames}
\title{Utilities for distributions supported in this package}
\usage{
check_distname(distname)

get_distname_family(distname)

get_distnames()
}
\arguments{
\item{distname}{string; name of input distribution; see \code{\link{get_distnames}}.}
}
\value{
\code{check_distname} returns (invisible) that the distribution is implemented,
or throws an error otherwise.

\code{get_distname_family} returns a list with
\item{location}{ logical; if \code{TRUE}, the distribution is a location family,}
\item{scale}{ logical; if \code{TRUE}, the distribution is a scale family.}

\code{get_distnames} returns a vector of strings in alphabetical order. It lists
all supported distributions.
Each string can be passed as the \code{distname} argument to several functions in this package.
}
\description{
The Lambert W\eqn{\times} F framework can take any (continuous) random variable with distribution
F and make it skewed (\code{type = "s"}), heavy tailed (\code{type = "h"}),
or both (\code{type = "hh"}).

In principle, this works for any F.  Of course, this package implements only a finite
number of distributions, which you usually pass with the \code{distname} argument.
Most functions in this package, however, also allow you to pass your own distribution and parameters
and create a Lambert W\eqn{\times} F version of it.
See \code{\link{create_LambertW_input}} and \code{\link{create_LambertW_output}}.

\code{check_distname} checks if the distribution specified by
the \code{distname} argument is implemented in this package.

\code{get_distname_family} determines whether a distribution is a location, scale, or
location-scale family.

\code{get_distnames} lists all currently implemented input distributions.
}
\examples{
check_distname("normal")
\dontrun{
check_distname("my_great_distribution")
}
get_distname_family("normal")
}
\keyword{misc}

