% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kurtosis}
\alias{kurtosis}
\alias{skewness}
\title{Skewness and kurtosis}
\usage{
kurtosis(x)

skewness(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
\code{kurtosis} estimates the fourth central, normalized moment from data.

\code{skewness} estimates the third central, normalized moment from data.
}
\seealso{
Corresponding functions in the \pkg{moments} package.
}
