% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_gamma.R
\name{H_gamma}
\alias{H_gamma}
\title{H transformation with gamma}
\usage{
H_gamma(u, gamma = 0)
}
\arguments{
\item{u}{a numeric vector of real values.}

\item{gamma}{skewness parameter; default \code{gamma = 0}, which implies
\code{H_gamma(u) = u}.}
}
\value{
numeric; same dimension/size as \code{u}
}
\description{
Skewed Lambert W\eqn{\times} F RV transformation: \eqn{H_{\gamma}(u) = u
    \exp(\gamma u)}.
}
\seealso{
\code{\link{xexp}}
}
\keyword{math}
