\name{LambertW-internal}
\alias{LambertW-internal}
\alias{mom.LambertW.U.Gauss}
\alias{mom.LambertW.X.Gauss}
\alias{sd}
\alias{MLE_LambertW_new}
\alias{MLE_LambertW_new.default}
\title{Internal LambertW objects}
\description{Internal LambertW objects.}
\details{These are not to be called by the user.}
\keyword{internal}
