\name{sd}
\alias{sd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Standard deviation}
\description{
The package \code{gsl} has one function called \code{sd} - get again the common \code{sd} function.
}
\usage{
sd(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ data }
}

\value{
empirical standard deviation
}
\references{Goerg, G.M. (2009). \dQuote{Lambert W Random Variables - A new class of skewed distribution functions}. Unpublished}
\author{Georg M. Goerg}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar}
