\name{vec.norm}
\alias{vec.norm}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Euclidean norm}
\description{
Calculates the Euclidean norm of  a vector.
}
\usage{
vec.norm(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ n-dimensional vector }
}

\value{
Euclidean norm of a vector \code{x} 
}

\author{Georg M. Goerg}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{math}
