\name{W}
\alias{W}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Lambert W function (principal branch) }
\description{
Principal branch of the Lambert W function. This is a short name for \code{lambert_W0}.
}
\usage{
W(z)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{z}{ value }
}

\value{
function value or NaN
}
\references{Goerg, G.M. (2009). \dQuote{Lambert W Random Variables - A new class of skewed distribution functions}. Unpublished}
\author{Georg M. Goerg}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{math}
