\name{H}
\alias{H}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Underlying transformation H}
\description{
 The Lambert W function is the inverse function of this function.
}
\usage{
H(u)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{u}{ value in R }
}

\value{
returns the value z = u exp(u).
}
\references{Goerg, G.M. (2009). \dQuote{Lambert W Random Variables - A new class of skewed distribution functions}. Unpublished}
\author{Georg M. Goerg}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{math}
