% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mc}
\alias{mc}
\title{MedCouple Estimator}
\usage{
mc(x)
}
\arguments{
\item{x}{numeric vector; if length > 3,000, it uses a random subsample (otherwise
\code{mc} takes too long to compute; \code{mc} calculations are of order \eqn{n^2}.)}
}
\value{
float; measures the degree of asymmetry
}
\description{
A robust measure of asymmetry. See References for details.
}
\examples{
# a little simulation/demonstration
kNumSim <- 100
kNumObs <- 200

################# Gaussian (Symmetric) ####
A <- t(replicate(kNumSim, {xx <- rnorm(kNumObs); c(skewness(xx), mc(xx))}))
########### skewed LambertW x Gaussian ####
tau.s <- gamma_01(0.2) # zero mean, unit variance, but positive skewness
rbind(mLambertW(theta = list(beta = tau.s[c("mu_x", "sigma_x")],
                             gamma = tau.s["gamma"]),
                distname="normal"))
B <- t(replicate(kNumSim,
                 {
                   xx <- rLambertW(n = kNumObs,
                                   theta = list(beta=tau.s[c("mu_x", "sigma_x")],
                                                gamma = tau.s["gamma"]),
                                   distname="normal")
                   c(skewness(xx), mc(xx))
                 }))

colnames(A) <- colnames(B) <- c("MedCouple", "Pearson Skewness")

layout(matrix(1:4, ncol = 2))
plot(A, main = "Gaussian")
boxplot(A)
abline(h = 0)

plot(B, main = "Skewed Lambert W x Gaussian")
boxplot(B)
abline(h = mLambertW(theta = list(beta = tau.s[c("mu_x", "sigma_x")],
                                  gamma = tau.s["gamma"]),
                     distname="normal")["skewness"])

colMeans(A)
apply(A, 2, sd)

colMeans(B)
apply(B, 2, sd)
}
\references{
Brys, G., M. Hubert, and A. Struyf (2004). \dQuote{A robust
measure of skewness}. Journal of Computational and Graphical Statistics 13
(4), 996 - 1017.
}
\keyword{univar}

