\name{sd}
\alias{sd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Standard deviation}
\description{
The package \code{gsl} has one function called \code{sd}, which does not calculate the sample standard deviation. Hence, here it is re-introduced as the original \code{stats:sd(x)} function.
}
\usage{
sd(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ data }
}

\value{
empirical standard deviation
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{univar}
