\name{support}
\alias{support}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Truncated support}
\description{
Given theta calculates the support of Y.
}
\usage{
support(theta)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theta}{ parameter vector }
}

\value{
half-open interval on the real line
}
\references{Goerg, G.M. (2009). \dQuote{Lambert W Random Variables - A new class of skewed distribution functions}. Unpublished}
\author{Georg M. Goerg}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{math}
