\name{W_1}
\alias{W_1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Lambert W function (non-principal branch)}
\description{
Non-principal solution of the Lambert W function. Short for \code{lambert_W1}.
}
\usage{
W_1(z)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{z}{value}
}

\value{
function value or NaN
}
\references{Goerg, G.M. (2009). \dQuote{Lambert W Random Variables - A new class of skewed distribution functions}. Unpublished}
\author{Georg M. Goerg}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{math}
