\name{W_delta}
\alias{W_delta}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Inverse transformation}
\description{
Inverse transformation of the Lambert W random variables. This function uses the principal branch.
}
\usage{
W_delta(z, delta)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{z}{ value }
  \item{delta}{ skewness parameter }
}

\value{
function value
}
\references{Goerg, G.M. (2009). \dQuote{Lambert W Random Variables - A new class of skewed distribution functions}. Unpublished}
\author{Georg M. Goerg}



% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{math}
